Conflicts:
	src/bms/player/beatoraja/play/BMSPlayer.java
	src/bms/player/beatoraja/play/LaneRenderer.java
	src/bms/player/beatoraja/play/PlaySkin.java

+++ src/bms/player/beatoraja/skin/SkinGraph.java
@@ -39,7 +39,7 @@ public class SkinGraph extends SkinObject {
 		if (getCycle() == 0) {
 			return image[0];
 		}
-		final int index = (int) ((time / (getCycle() / image.length))) % image.length;
+		final int index = ((int) (time / (((float)getCycle())  / image.length))) % image.length;
 		// System.out.println(index + " / " + image.length);
 		return image[index];
 	}
